set panda display every two hours
