+++ front/src/components/templates/Footer.styled.tsx
@@ -1,10 +1,8 @@
-/* eslint-disable no-confusing-arrow */
-/* eslint-disable object-curly-newline */
-import React, { useContext, useEffect, useRef, useState } from 'react';
+import React, { useContext } from 'react';
 import styled, { ThemeContext } from 'styled-components';
 import ColorLuminance from '../../utils/colorLuminance';
-import Tabable from '../core/accessibility/Tabable.Styled';
 import LargeDivider from '../core/LargeDivider.styled';
+import PandaTips from './PandaTips.styled';
 
 /**
  * Builds a responsive footer with a neumorphic divider on top
@@ -39,18 +37,6 @@ const Wrapper = styled.div`
 	}
 `;
 
-const Image = styled.img`
-	width: 100px;
-	position: absolute;
-	bottom: 100px;
-	right: 100px;
-	cursor: pointer;
-`;
-
-const Timer = styled.div<{ visible?: boolean }>`
-	visibility: ${(props) => props.visible ? 'visible' : 'hidden'};
-`;
-
 // TODO: make a clickable link and create view for terms of services.
 const Footer = (): JSX.Element => {
 	/// Compute shadows colors from current primary theme color
@@ -59,36 +45,9 @@ const Footer = (): JSX.Element => {
 	const lightShadow = ColorLuminance(primaryColor, -0.5);
 	const darkShadow = ColorLuminance(primaryColor, +0.5);
 
-	const [pandaVisible, setPandaVisibility] = useState(false);
-
-	useEffect(() => {
-		// Show panda every 2 hours
-		const timer = setInterval(() => {
-			console.log('This will pop every 2 hours');
-			setPandaVisibility(true);
-			// Hide panda after 5 minutes
-			setTimeout(() => { setPandaVisibility(false); }, 300000);
-		}, 12000); //	7200000
-		return () => clearInterval(timer);
-	}, []);
-
 	return (
 		<FooterContainer>
-			<Timer visible={pandaVisible}>
-				<Tabable
-					onClick={() => {
-						console.log('clicked !');
-						setPandaVisibility(false);
-					}}
-					pressedKey="Enter"
-					onKeyPress={() => {
-						console.log('key pressed !');
-						setPandaVisibility(false);
-					}}
-				>
-					<Image src="/assets/images/133706.png" alt="open panda tips" />
-				</Tabable>
-			</Timer>
+			<PandaTips />
 			<LargeDivider light={lightShadow} dark={darkShadow} />
 			<FooterContent data-testid="footer">
 				<Wrapper>© 2021 Alpha Wilders - All Rights Reserved.</Wrapper>
